Add `stack_deployed_by` and update stack version handling (#5)
* Add stack_deployed_by and update stack version handling

Adds new stack_deployed_by Agent Metadata tags to properly identify which IaC method was used to deploy the Stack. As such, this also hardcodes the value "Terraform" in the stack when passing the variable to the user-data templates.

Aligns with the CloudFormation equivalent in values passed to Agent Metadata.

No further changes needed to align with v6.44.0 (when released) as changes were all Packer based. Will still need to update `cloudformation_stack_version` var, and AMI's before raising PR.

* Bump AMI versions and CF stack version

* Terraform fmt

* T > t