Watching CFT version now, too.
This is gonna be a bit mad, but gonna go ahead and ingest this in to CI and watch the cft template for an update, once it matches the version proposed, it will go ahead and push an update to the locals. Just gonna sort a local script first that can do that.