feat: Remove Scaler var, update with renovate (#30)
This PR replicates https://github.com/buildkite/elastic-ci-stack-for-aws/pull/1691 which removes the Scaler Version as a a configurable.

This will be set via local now and also watched using Renovate to update.

This PR also introduces some improvements such as updating versions for the module in README.md and all main.tf's in the examples.

This will also update all provider versions in versions.tf

Additionally, due to the Renovate changes, we now have a pipeline to watch for changes, if a change to the ElasticStack version takes place, AMI mappings will automatically be updated and a comment will be left to advise of this, as well as a reminder to check for any changes such as variable changes, etc before merging. This can be seen from within this PR working.

For the time being, we're just adding packages to the terraform base image, once available, this will be moved to a dedicated Dockerfile, but I'm unsure how long I'll be blocked on that, so pressing on as-is as this is working.

I dropped versions of the module to non-current versions (scaler too)